Необходимые правки для решения задания 8-2
